+++ src/lib.rs
@@ -1,11 +1,11 @@
 #![allow(non_snake_case)]
+#![allow(non_camel_case_types)]
 pub mod constants;
 pub mod cosmology;
 pub mod dark_energy;
 pub mod distances;
 pub mod redshift;
 pub mod units;
-pub mod utils;
 
 pub use cosmology::FLRWCosmology;
 pub use distances::Distances;
@@ -13,6 +13,14 @@ pub use distances::Distances;
 // Common units are re-exported from the crate root for convenience.
 pub use redshift::Redshift;
 pub use units::{
-    eV, DimensionlessFloat, DimensionlessPositiveFloat, HInvKmPerSecPerMpc, Joule, Kelvin,
-    Kilogram, KmPerSecPerMpc, Mpc,
+    energy::{eV, Joule},
+    length::{Kilometer, Meter, Mpc},
+    mass::{Gram, Kilogram},
+    temperature::Kelvin,
+    time::{Gyr, Seconds},
+    DimensionlessFloat, DimensionlessPositiveFloat, HInvKmPerSecPerMpc, KilogramsPerMeter3,
+    KmPerSecPerMpc,
 };
+
+// Common traits are re-exported from the crate root also.
+pub use crate::units::traits::FloatingPointUnit;