+++ src/units.rs
@@ -1,32 +1,23 @@
 use anyhow::anyhow;
 
+pub mod dimensionless;
+pub mod energy;
+pub mod length;
+pub(crate) mod macros;
+pub mod mass;
+pub mod temperature;
+pub mod time;
+pub mod traits;
+
+pub use dimensionless::DimensionlessFloat;
+pub use traits::FloatingPointUnit;
+
 // Continuous positive quantities that are dimensionless (e.g. ratios like the omegas)
 pub type DimensionlessPositiveFloat = PositiveFloat;
-
-// Time
-pub type Gyr = PositiveFloat;
-pub type Seconds = PositiveFloat;
-
-// Energy
-#[allow(non_camel_case_types)]
-pub type eV = PositiveFloat;
-pub type Joule = PositiveFloat;
-
-// Mass
-pub type Kilogram = PositiveFloat;
-pub type Gram = PositiveFloat;
-
-// Temperatures
-pub type Kelvin = PositiveFloat;
 
 // Hubble parameter units
 pub type KmPerSecPerMpc = f64;
 pub type HInvKmPerSecPerMpc = f64;
-
-// Distances
-pub type Meter = PositiveFloat;
-pub type Km = PositiveFloat;
-pub type Mpc = PositiveFloat;
 
 // Densities
 pub type KilogramsPerMeter3 = PositiveFloat;
@@ -40,54 +31,8 @@ pub type JouleSeconds = f64;
 pub type JoulePerMeter3Kelvin4 = f64;
 pub type WattsPerMeters2Kelvin4 = f64;
 pub type JoulePerKelvin = f64;
-
-// Conversions
-pub const KILOMETER_TO_METER: f64 = 1000.;
-pub const MPC_TO_METERS: f64 = 3.086e+22;
-pub const MPC_TO_KILOMETERS: f64 = 3.086e+19;
-
-/// Represents continuous dimensionless physical quantities.
-#[derive(Clone, Copy, Debug, PartialOrd, PartialEq)]
-pub struct DimensionlessFloat(pub f64);
-
-impl DimensionlessFloat {
-    pub fn new(x: f64) -> Result<Self, anyhow::Error> {
-        Ok(Self(x))
-    }
-
-    pub fn zero() -> Self {
-        Self(0.)
-    }
-
-    pub fn one() -> Self {
-        Self(1.)
-    }
-
-    // Passthrough methods for convenience
-    pub fn floor(&self) -> f64 {
-        self.0.floor()
-    }
-
-    pub fn powf(&self, exp: f64) -> f64 {
-        self.0.powf(exp)
-    }
-}
-
-impl std::ops::Sub for DimensionlessFloat {
-    type Output = DimensionlessFloat;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        DimensionlessFloat(self.0 - rhs.0)
-    }
-}
-
-impl std::ops::Add for DimensionlessFloat {
-    type Output = DimensionlessFloat;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        DimensionlessFloat(self.0 + rhs.0)
-    }
-}
+pub type Mpc3 = f64;
+pub type HInvMpc = f64;
 
 /// Represents continuous physical quantities that _cannot_ be negative.
 #[derive(Clone, Copy, Debug, PartialOrd, PartialEq)]