+++ src/units/dimensionless.rs
@@ -2,4 +2,4 @@ use std::ops::{Add, Sub};
 
 use crate::units::{macros::floating_point_unit_impl, traits::FloatingPointUnit};
 
-floating_point_unit_impl! { Redshift }
+floating_point_unit_impl! { DimensionlessFloat }