+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.2.0
+
+* Adds lookback time and distance calculations.
+* Adds comoving volume calculations.
+* Updates values for constants based on CODATA 2018.
+* Refactors handling of base units.
+* Add initial benchmarks.
+* Performance improvements.
+
 # 0.1.0
 
 Initial release.